+++ Web/src/main/java/com/github/reloadedd/correctanaddress/routes/CorrectAddressRoute.java
@@ -18,13 +18,19 @@
 
 package com.github.reloadedd.correctanaddress.routes;
 
-import com.github.reloadedd.correctanaddress.entities.AddressEntity;
-import com.github.reloadedd.correctanaddress.entities.GoogleMapsQueryEntity;
-import org.springframework.stereotype.Controller;
+import java.util.Collections;
+import org.springframework.http.*;
 import org.springframework.ui.Model;
+import org.springframework.util.MultiValueMap;
+import org.springframework.stereotype.Controller;
+import org.springframework.web.client.RestTemplate;
+import org.springframework.util.LinkedMultiValueMap;
 import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.boot.web.client.RestTemplateBuilder;
+import com.github.reloadedd.correctanaddress.entities.GoogleMapsQueryEntity;
+import com.github.reloadedd.correctanaddress.entities.HereMapsAddressEntity;
 
 
 /**
@@ -36,13 +42,29 @@ import org.springframework.web.bind.annotation.RequestParam;
 @Controller
 @RequestMapping("/correct-this-address")
 public class CorrectAddressRoute {
+    private static final String CORRECT_ADDRESS_API_URL = "https://reloadedd.me:5443/api/correct-this-address";
+
     @PostMapping
     public String getCorrectAddress(@RequestParam String country, @RequestParam String state,
                                     @RequestParam String city, Model model) {
-        AddressEntity addressEntity = new AddressEntity(country, state, city);
-        model.addAttribute("address", addressEntity.constructAddress());
+        final RestTemplate restTemplate = new RestTemplateBuilder().build();
+        MultiValueMap<String, String> map= new LinkedMultiValueMap<>();
+        map.put("country", Collections.singletonList(country));
+        map.put("state", Collections.singletonList(state));
+        map.put("city", Collections.singletonList(city));
+
+        /* Set Content-Type header to JSON */
+        HttpHeaders headers = new HttpHeaders();
+        headers.setContentType(MediaType.APPLICATION_FORM_URLENCODED);
+
+        HttpEntity<MultiValueMap<String, String>> request = new HttpEntity<>(map, headers);
+        HereMapsAddressEntity response = restTemplate.postForObject(CORRECT_ADDRESS_API_URL, request,
+                HereMapsAddressEntity.class);
+
+        assert response != null;
+        model.addAttribute("address", response.constructAddress());
         model.addAttribute("query",
-                new GoogleMapsQueryEntity(addressEntity).getGoogleMapsURL());
+                new GoogleMapsQueryEntity(response.getGoogleMapsAddress()).getGoogleMapsURL());
 
         return "results";
     }