+++ CHANGELOG.md
@@ -5,10 +5,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-## Added
+## [0.6.0] - 2021-05-25
+### Added
 - `/results` route that will display the corrected address into a Google Maps view.
+- The REST API is working and corrects addresses based on the query score received from the
+HERE Maps API.
 
-## Modified
+### Modified
 - Move `Jenkinksfile` to the root
 - Modify `Jenkinsfile` to take into account the new folder structure